Use paths relative to current working directory
Previously, using the construct `__DIR__ . '/../data/' for paths to keys
causes fatal errors immediately when the user does not override the
configuration, and instead uses the default configuration. This is
particularly problematic because the documentation suggests that running
`vendor/bin/generate-oauth2-keys` should be sufficient.

This patch modifies those paths to be relative to the current working
directory, `getcwd() . '/data'`, fixing those issues.